// src/app/model/registries/solutions-info.ts
import {BaseInfo} from '../BaseInfo';

export class SolutionsInfo extends BaseInfo {
    Name: string;
    One: number;
    Two: number;
    Three: number;
    Four: number;
    Five: number;
    Six: number;
    Seven: number;
    Eight: number;
    Nine: number;
    Ten: number;
    Eleven: number;
}